% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{baseLaTeXHighlighter}
\alias{baseLaTeXHighlighter}
\title{Base LaTeX highlighter}
\usage{
baseLaTeXHighlighter(values, ordering, n.highlight = 1,
  bg.color = "gray", bg.saturation.max = 80, bg.saturation.min = 20)
}
\arguments{
\item{values}{[\code{numeric}]\cr
Vector of numeric values.}

\item{ordering}{[\code{integer}]\cr
Ordering of \code{values}. I.e., if \code{ordering[i] == 3}, then
the i-th element is 3rd in the sorted sequence.}

\item{n.highlight}{[\code{integer(1)}]\cr
The \code{n.highlight} best -- according to \code{ordering} --
values are highlighted with a background color.}

\item{bg.color}{[\code{character(1)}]\cr
Base color used for LaTeX table cellcolor.}

\item{bg.saturation.max}{[\code{character(1)}]\cr
Maximal color saturation used for cellcolor.}

\item{bg.saturation.min}{[\code{character(1)}]\cr
Minimal color saturation used for cellcolor.}
}
\value{
[\code{character}] Modified \code{values} vector.
}
\description{
Highlight best values and optionally highlights the
\code{k} best values with background colors of different opacity.
}
