% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_convergence_plot.R
\name{convergence_plot}
\alias{convergence_plot}
\title{Convergence plot}
\usage{
convergence_plot(tbl, x, y, color = NULL, linetype = color,
  shape = NULL, xlab = x, ylab = y, vlines.at = NULL,
  hlines.at = NULL, x.logscale.base = NULL, ...)
}
\arguments{
\item{tbl}{[\code{data.frame}]\cr
Data frame with input data.}

\item{x}{[\code{string}]\cr
Column name of x-value.}

\item{y}{[\code{string}]\cr
Column name of y-value.}

\item{color}{[\code{string}]\cr
Column name of column used for coloring.}

\item{linetype}{[\code{string}]\cr
Column name of column used for linetype. Defaults to \code{color}.}

\item{shape}{[\code{string}]\cr
Column name of column used for shape. If not \code{NULL}, in addition to lines,
points are plotted.}

\item{xlab}{[\code{string}]\cr
x-axis label. Defaults to \code{x}.}

\item{ylab}{[\code{string}]\cr
y-axis label. Defaults to \code{y}.}

\item{vlines.at}{[\code{numeric}]\cr
If not \code{NULL}, indicates the positions of dashed gray vertical lines.}

\item{hlines.at}{[\code{numeric}]\cr
If not \code{NULL}, indicates the positions of dashed gray horizontal lines.}

\item{x.logscale.base}{[\code{numeric(1)}]\cr
Base used for log-scaled abszissa.}

\item{...}{[any]\cr
Not used.}
}
\value{
[\code{\link[ggplot2]{ggplot}}]
}
\description{
Line plot of trajectories. Given a data frame and values for
abszissa and ordinate the function returns a modifiable ggplot2 object.
}
