% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview}
\alias{preview}
\title{Preview of LaTeX tables.}
\usage{
preview(...)
}
\arguments{
\item{...}{[any]\cr
list of kable tables.}
}
\value{
[invisible(NULL)]
}
\description{
Preview of LaTeX tables.
}
