% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_result_table.R
\name{to_latex}
\alias{to_latex}
\title{Convert result table into basic kable LaTeX-table.}
\usage{
to_latex(tbl, param.col.names, measure.col.names, reps = 1L, algo.names,
  algo.colors = NULL, caption = "Placeholder", longtable = FALSE)
}
\arguments{
\item{tbl}{[|code{data.frame}]\cr
Input data frame (usually result of \code{\link{to_result_table}}).}

\item{param.col.names}{[\code{character}]\cr
Names of parameter columns used for table header. Parameter column names
are supposed to be the first \eqn{k} column in \code{tbl}.}

\item{measure.col.names}{[\code{character}]\cr
Names of measurement columns. These are repeated for each algorithm (see
\code{algo.names}).}

\item{reps}{[\code{integer(1)}]\cr
How often should measurement columns be repeated? Defaults to 1.
This is useful if results for multiple performance measures should be displayed
side by side in the same table.}

\item{algo.names}{[\code{character}]\cr
Names of algorithms. Used to group measurement columns. Note that algorithms
are numbered by their index.}

\item{algo.colors}{[\code{character}]\cr
Vector of color names used to color-code algorithms for better visual
interpretation.}

\item{caption}{[\code{character(1)}]\cr
LaTeX-table caption.}

\item{longtable}{[|code{logical(1)}]\cr
Should the table be used in a longtable environment?
Default is \code{FALSE}.}
}
\value{
[character(1)] A character vector of the table source code.
}
\description{
Used to generate a basic result table object based on results
of \code{to_result_table}.
}
