% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_result_table.R
\name{to_result_table}
\alias{to_result_table}
\title{Generate data frame with aggregated data and statistical test results.}
\usage{
to_result_table(x, split.cols, widen.col, measure.cols,
  do.pairwise.test = TRUE, test.alternative,
  testresult.col = "testresult", alpha = 0.05,
  stats.formatter.args = list(sd = list(digits = 2L)),
  testresult.formatter.args = list(positive.only = TRUE),
  highlighter.args = NULL)
}
\arguments{
\item{x}{[\code{data.frame}]\cr
Input data frame.}

\item{split.cols}{[\code{character}]\cr
Column name of columns used to split the data. Usually this will be something
like \dQuote{problem/instance}, \dQuote{parameter}.}

\item{widen.col}{[\code{character(1)}]\cr
Column used to widen the table, convert long to wide format. Usually, this is
the \dQuote{algorithm}.}

\item{measure.cols}{[\code{character}]\cr
Which columns contain the actual measurements (e.g. performance values, number
of iterations etc.). If multiple are provided and \code{do.pairwise.test} is
\code{TRUE}, the data is additionally split by \code{measure.cols}.}

\item{do.pairwise.test}{[\code{logical(1)}]\cr
Should pairwise test of measurements be performed?
Default is \code{TRUE}.}

\item{test.alternative}{[\code{character}]\cr
Alternatives, i.e. \eqn{H_1}-hypothesis, for statistical tests.}

\item{testresult.col}{[\code{character}]\cr
Column name where test result shall be stored.
Default is \dQuote{testresult}.}

\item{alpha}{[\code{numeric(1)}]\cr
Significance level for statistical tests. Default to \code{0.05}.}

\item{stats.formatter.args}{[\code{list of lists}]\cr
Named list (names are aggregated measures) of arguments used to format the
aggregated measurement.}

\item{testresult.formatter.args}{[\code{list}]\cr
Named list of parameters passed down to LaTeX-formatter for test results.}

\item{highlighter.pars}{[\code{list}]\cr
Named list of arguments for highlighter, i.e. function that is used to
emphasize certain results in tables (e.g. the minimal mean value on each
instance).}
}
\value{
[\code{list}] List of aggregated data frames.
}
\description{
Often we are given a large \code{data.frame} with results of
multiple (stochastic/randomized) algorithms on multiple instances with varying parameters.
One common way to present such results is to aggregate over runs, e.g. on instance
level, apply statistical singnificance tests to underpin observations and
eventually to present these ggregated results by means of result tables.
These result tables contain meta columns (e.g. problem, parameters) and each
at least one column for each algorithm.

This function serves for aggregation and calculation of statistical tests.
The result is a list of aggregated result tables which may be passed on to
\code{\link{to_latex}}.
}
