% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widen.R
\name{widen}
\alias{widen}
\title{Widen data frame}
\usage{
widen(x, split.col, widen.cols)
}
\arguments{
\item{x}{[\code{data.frame}]\cr
Data frame.}

\item{split.col}{[\code{character(1)}]\cr
Column name of column used for splitting.}

\item{widen.cols}{[\code{character}]\cr
Column name(s) of column(s) which shall be aligned beside
each other.}
}
\value{
[\code{data.frame}]
}
\description{
Split a data frame by a factor or character
column and piece together a wide data frame by arranging
socalled widening columns side by side.
}
