% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.pairwise.R
\name{test.pairwise}
\alias{test.pairwise}
\title{Appnds formatted results of pairwise tests to data frame}
\usage{
test.pairwise(x, by, split.col, value.col, testresult.col, alpha = 0.05,
  test.fun = stats::wilcox.test, ...)
}
\arguments{
\item{x}{[\code{data.frame}]\cr
Data frame.}

\item{by}{[\code{character}]\cr
Column names used for splitting the data frame.
This will usually be columns that describe the problems / problem parameters.}

\item{split.col}{[\code{character(1)}]\cr
Single column name. This will usually be the algorithm column.}

\item{value.col}{[\code{character(1)}]\cr
Single column name with numeric values to undergo statistical tests.}

\item{testresult.col}{[\code{character(1)}]\cr
Column name of new column where test results should be stored.}

\item{alpha}{[\code{numeric(1)}]\cr
Significance level.
Default is 0.05, i.e. five percent.}

\item{test.fun}{[\code{function(x, y, ...)}]\cr
Statistical test function.
Default is \code{wilcox.test}.}

\item{...}{[any]\cr
Further parameters passed down to \code{test.fun}.
E.g., \code{alternative = "less"}.}
}
\value{
[\code{data.frame}]
}
\description{
Split by some columns, further split by \dQuote{split.col},
and do pairwise comparisons of the \code{value.col} column based on
a statistical test. Results are stored in \code{testresult.col}.
}
